scheduler: move into internal/, split into files
Move scheduler.go into internal/scheduler,
rename and export things,
and split the implementations across files.
